Linux builds now uses tshark over jnetpcap.
To install tshark: `sudo apt-get install tshark`.

Although jnetpcap is open source, it is not free software. A purchased license is required for
commercial use. To allow EnergyBox builds to be used for commercial purposes this commit
sets the Linux build to use tshark (which was used by OSX previously). Windows builds
still use jnetpcap and should thus be built with `gradle jnetpcapJar`.

- New build target "jnetpcapJar" builds a jar that includes jnetpcap library
  - Usage: `gradle jnetpcapJar`
- Rename ProcessTraceOSX --> ProcessTraceTshark to better reflect its functionality